separated the roi_processing functions into a separate file
